+++ extensions/pieces-raycast/src/connection/health/piecesUpToDateCheck.ts
@@ -1,13 +1,14 @@
 import * as semver from "semver";
 import ConnectorSingleton from "../ConnectorSingleton";
-import { showToast, open, Toast } from "@raycast/api";
+import { showToast, Toast } from "@raycast/api";
 import Notifications from "../../ui/Notifications";
 import { UpdatingStatusEnum } from "@pieces.app/pieces-os-client";
 import { pollForConnection } from "./piecesHealthCheck";
 import sleep from "../../utils/sleep";
+import BrowserUrl from "../../utils/BrowserUrl";
 
-const MIN_VERSION = "11.0.0";
-const MAX_VERSION = "12.0.0";
+const MIN_VERSION = "12.0.0";
+const MAX_VERSION = "13.0.0";
 
 /**
  * Checks if the Pieces version is up to date and handles necessary updates.
@@ -72,7 +73,7 @@ async function updatePieces(): Promise<boolean> {
         toast.primaryAction = {
           title: "Contact Support",
           onAction() {
-            open("https://docs.pieces.app/support");
+            BrowserUrl.open("https://docs.pieces.app/products/support");
           },
         };
       }
@@ -106,7 +107,7 @@ function getStatusText(status: UpdatingStatusEnum | undefined) {
     case UpdatingStatusEnum.Available:
       return "Update detected...";
     case UpdatingStatusEnum.ContactSupport:
-      return "Something went wrong. Please contact support at https://docs.pieces.app/support";
+      return "Something went wrong. Please contact support at https://docs.pieces.app/products/support";
     case UpdatingStatusEnum.Downloading:
       return "Update is downloading...";
     case UpdatingStatusEnum.ReadyToRestart:
@@ -118,7 +119,7 @@ function getStatusText(status: UpdatingStatusEnum | undefined) {
     case UpdatingStatusEnum.UpToDate:
       return "PiecesOS is up to date.";
     case undefined:
-      return "Failed to get update status, please contact support at https://docs.pieces.app/support";
+      return "Failed to get update status, please contact support at https://docs.pieces.app/products/support";
   }
 }